requirements.txt
numpy
pyqt5
pyopengl
